some code quality and consistency improvements for quizzes

+++ src/main/java/de/tum/in/www1/artemis/domain/quiz/QuizExercise.java
@@ -339,7 +339,7 @@ public class QuizExercise extends Exercise implements Serializable {
      */
     public Long getScoreForSubmission(QuizSubmission quizSubmission) {
         double score = getScoreInPointsForSubmission(quizSubmission);
-        int maxScore = getMaxTotalScore();
+        double maxScore = getMaxTotalScore();
         // map the resulting score to the 0 to 100 scale
         return Math.round(100.0 * score / maxScore);
     }
@@ -525,8 +525,8 @@ public class QuizExercise extends Exercise implements Serializable {
      * @return the sum of all the quizQuestions' maximum scores
      */
     @JsonIgnore
-    public Integer getMaxTotalScore() {
-        int maxScore = 0;
+    public Double getMaxTotalScore() {
+        double maxScore = 0.0;
         // iterate through all quizQuestions of this quiz and add up the score
         if (quizQuestions != null && Hibernate.isInitialized(quizQuestions)) {
             for (QuizQuestion quizQuestion : getQuizQuestions()) {
@@ -544,7 +544,7 @@ public class QuizExercise extends Exercise implements Serializable {
             return score;
         }
         else if (quizQuestions != null && Hibernate.isInitialized(quizQuestions)) {
-            return getMaxTotalScore().doubleValue();
+            return getMaxTotalScore();
         }
         return null;
     }
